requirements.txt
fastapi==0.95.1
ldap3==2.9.1
pyJWT==2.6.0
uvicorn==0.22.0
python-multipart==0.0.6
